refactor(web): split agents page sections
